python-ai/jakhangirs-core/requirements.txt
openai>=1.66.0
anthropic>=0.49.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
